server/app/routers/kernel.py
"""
Semantic Kernel Management Router

This router provides endpoints for managing Semantic Kernel plugins
and demonstrating advanced AI orchestration capabilities.
"""

import json
import logging
from fastapi import APIRouter, HTTPException, Request
from fastapi.responses import JSONResponse
from typing import Dict, Any, Optional
from pydantic import BaseModel

from app.kernel.services.enhanced_kernel_service import get_enhanced_kernel_service

# Configure logging
log = logging.getLogger("kernel_router")

router = APIRouter(prefix="/api/kernel", tags=["semantic-kernel"])

class PluginStatusResponse(BaseModel):
    """Response model for plugin status."""
    plugin_name: str
    status: str
    functions: list

@router.get("/status")
async def get_kernel_status():
    """
    Get basic status of the Semantic Kernel.
    
    Returns:
        Basic kernel health and plugin information
    """
    try:
        enhanced_service = get_enhanced_kernel_service()
        is_healthy = await enhanced_service.validate_health()
        
        return JSONResponse({
            "kernel_initialized": True,
            "healthy": is_healthy,
            "plugins_count": len(enhanced_service.plugins),
            "services_available": len(enhanced_service.kernel.services) > 0
        })
        
    except Exception as exc:
        log.error("Failed to get kernel status: %s", exc)
        raise HTTPException(status_code=500, detail=f"Failed to get kernel status: {exc}")

@router.get("/health")
async def kernel_health_check():
    """
    Perform a comprehensive health check of the Semantic Kernel.
    
    Returns:
        Health status and diagnostic information
    """
    try:
        enhanced_service = get_enhanced_kernel_service()
        is_healthy = await enhanced_service.validate_health()
        
        return JSONResponse({
            "healthy": is_healthy,
            "status": "healthy" if is_healthy else "unhealthy",
            "timestamp": "2025-07-15T00:00:00Z",  # Current timestamp would be generated
            "plugins_count": len(enhanced_service.plugins)
        })
        
    except Exception as exc:
        log.error("Kernel health check failed: %s", exc)
        return JSONResponse(
            status_code=500,
            content={
                "healthy": False,
                "status": "error",
                "error": str(exc),
                "timestamp": "2025-07-15T00:00:00Z"
            }
        )
