server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai>=1.12,<2.0
azure-cosmos
semantic-kernel>=1.0.0
azure-identity>=1.15.0
